Vipin

+++ src/controller/category.controller.js
@@ -1,6 +1,8 @@
 const catchAsyncErrors = require("../middleware/catchAsyncErrors");
 const ErrorHandler = require("../utils/errorHandler");
 const categoryService = require("../services/category.service");
+const { cloudinary } = require("../utils/cloudinary");
+const Image = require("../model/image.model");
 
 // create category
 exports.createCategory = catchAsyncErrors(async (req, res, next) => {
@@ -65,12 +67,31 @@ exports.getCategoryById = catchAsyncErrors(async (req, res, next) => {
 
 exports.updateCategory = catchAsyncErrors(async (req, res, next) => {
   const { name, description, image } = req.body;
-  const category = await categoryService.updateCategory(req.params.id, { name, description, image });
+  const categoryData = await categoryService.getCategoryById(req.params.id);
+  if (!categoryData) return next(new ErrorHandler("No category found", 404));
+  let category
+  
+  const imageId = categoryData.image?._id?.toString();
+
+  if (imageId && imageId !== image) {
+    await cloudinary.uploader.destroy(categoryData.image.public_id);
+    await Image.deleteOne({ _id: categoryData.image._id });
+    category = await categoryService.updateCategory(req.params.id, { name, description, image });
+  } else {
+    category = await categoryService.updateCategory(req.params.id, { name, description,image });
+  }
+  
   if (!category) return next(new ErrorHandler("No category found", 404));
   res.status(200).json({ success: true, category, message: "category updated successfully" });
 });
 
 exports.deleteCategory = catchAsyncErrors(async (req, res, next) => {
+  const categoryData = await categoryService.getCategoryById(req.params.id);
+  if (!categoryData) return next(new ErrorHandler("No category found", 404));
+  if(categoryData.image){
+    await cloudinary.uploader.destroy(categoryData.image.public_id);
+    await Image.deleteOne({ _id: categoryData.image._id });
+  }
   const category = await categoryService.deleteCategory(req.params.id);
   if (!category) return next(new ErrorHandler("No category found", 404));
   res.status(200).json({ success: true, category, message: "category deleted successfully" });
